refactor code into table_monitor.py
